LayoutTests/editing/deleting/delete-emoji-9.html

<!DOCTYPE html>
<html>
<body>
<div id="test" contenteditable="true">
&#x1F919;&#x1F3FB;&#x1F468;&zwj;&#x2695;&#xFE0F;&#x1F468;&#x1F3FB;&zwj;&#x2695;&#xFE0F;&#x1F468;&zwj;&#x2695;&#x1F468;&#x1F3FB;&zwj;&#x2695;&#x1f3f3;&#xfe0f;&#x200d;&#x1f307;&#x1f3f3;&#xfe0f;&#x200d;&#x1f306;&#x1f3f3;&#xfe0f;&#x200d;&#x1f305;&#x1f3f4;&#xe0067;&#xe0062;&#xe0065;&#xe006e;&#xe0067;&#xe007f;&#x1f3f4;&#xe0067;&#xe0062;&#xe0073;&#xe0063;&#xe0074;&#xe007f;&#x1f3f4;&#xe0067;&#xe0062;&#xe0077;&#xe006c;&#xe0073;&#xe007f;</div>
<script src="../../resources/dump-as-markup.js"></script>
<script>
Markup.description("This test verifies that deletions are correct over emoji groups and emoji with variations");
var testElement = document.getElementById('test');
getSelection().setBaseAndExtent(testElement.firstChild, testElement.firstChild.length, testElement.firstChild, testElement.firstChild.length);
Markup.dump("test");
while (testElement.firstChild.length) {
    document.execCommand("Delete");
    Markup.dump("test");
}

</script>
</body>
</html>
