LayoutTests/platform/mac-sierra/editing/deleting/delete-emoji-9-expected.txt

This test verifies that deletions are correct over emoji groups and emoji with variations

Dump of markup 1:
| "
🤙🏻👨‍⚕️👨🏻‍⚕️👨‍⚕👨🏻‍⚕🏳️‍🌇🏳️‍🌆🏳️‍🌅🏴󠁧󠁢󠁥󠁮󠁧󠁿🏴󠁧󠁢󠁳󠁣󠁴󠁿🏴󠁧󠁢󠁷󠁬󠁳󠁿<#selection-caret>
"

Dump of markup 2:
| "
🤙🏻👨‍⚕️👨🏻‍⚕️👨‍⚕👨🏻‍⚕🏳️‍🌇🏳️‍🌆🏳️‍🌅🏴󠁧󠁢󠁥󠁮󠁧󠁿🏴󠁧󠁢󠁳󠁣󠁴󠁿🏴󠁧󠁢󠁷󠁬󠁳<#selection-caret>"

Dump of markup 3:
| "
🤙🏻👨‍⚕️👨🏻‍⚕️👨‍⚕👨🏻‍⚕🏳️‍🌇🏳️‍🌆🏳️‍🌅🏴󠁧󠁢󠁥󠁮󠁧󠁿🏴󠁧󠁢󠁳󠁣󠁴󠁿🏴󠁧󠁢󠁷󠁬<#selection-caret>"

Dump of markup 4:
| "
🤙🏻👨‍⚕️👨🏻‍⚕️👨‍⚕👨🏻‍⚕🏳️‍🌇🏳️‍🌆🏳️‍🌅🏴󠁧󠁢󠁥󠁮󠁧󠁿🏴󠁧󠁢󠁳󠁣󠁴󠁿🏴󠁧󠁢󠁷<#selection-caret>"

Dump of markup 5:
| "
🤙🏻👨‍⚕️👨🏻‍⚕️👨‍⚕👨🏻‍⚕🏳️‍🌇🏳️‍🌆🏳️‍🌅🏴󠁧󠁢󠁥󠁮󠁧󠁿🏴󠁧󠁢󠁳󠁣󠁴󠁿🏴󠁧󠁢<#selection-caret>"

Dump of markup 6:
| "
🤙🏻👨‍⚕️👨🏻‍⚕️👨‍⚕👨🏻‍⚕🏳️‍🌇🏳️‍🌆🏳️‍🌅🏴󠁧󠁢󠁥󠁮󠁧󠁿🏴󠁧󠁢󠁳󠁣󠁴󠁿🏴󠁧<#selection-caret>"

Dump of markup 7:
| "
🤙🏻👨‍⚕️👨🏻‍⚕️👨‍⚕👨🏻‍⚕🏳️‍🌇🏳️‍🌆🏳️‍🌅🏴󠁧󠁢󠁥󠁮󠁧󠁿🏴󠁧󠁢󠁳󠁣󠁴󠁿🏴<#selection-caret>"

Dump of markup 8:
| "
🤙🏻👨‍⚕️👨🏻‍⚕️👨‍⚕👨🏻‍⚕🏳️‍🌇🏳️‍🌆🏳️‍🌅🏴󠁧󠁢󠁥󠁮󠁧󠁿🏴󠁧󠁢󠁳󠁣󠁴󠁿<#selection-caret>"

Dump of markup 9:
| "
🤙🏻👨‍⚕️👨🏻‍⚕️👨‍⚕👨🏻‍⚕🏳️‍🌇🏳️‍🌆🏳️‍🌅🏴󠁧󠁢󠁥󠁮󠁧󠁿🏴󠁧󠁢󠁳󠁣󠁴<#selection-caret>"

Dump of markup 10:
| "
🤙🏻👨‍⚕️👨🏻‍⚕️👨‍⚕👨🏻‍⚕🏳️‍🌇🏳️‍🌆🏳️‍🌅🏴󠁧󠁢󠁥󠁮󠁧󠁿🏴󠁧󠁢󠁳󠁣<#selection-caret>"

Dump of markup 11:
| "
🤙🏻👨‍⚕️👨🏻‍⚕️👨‍⚕👨🏻‍⚕🏳️‍🌇🏳️‍🌆🏳️‍🌅🏴󠁧󠁢󠁥󠁮󠁧󠁿🏴󠁧󠁢󠁳<#selection-caret>"

Dump of markup 12:
| "
🤙🏻👨‍⚕️👨🏻‍⚕️👨‍⚕👨🏻‍⚕🏳️‍🌇🏳️‍🌆🏳️‍🌅🏴󠁧󠁢󠁥󠁮󠁧󠁿🏴󠁧󠁢<#selection-caret>"

Dump of markup 13:
| "
🤙🏻👨‍⚕️👨🏻‍⚕️👨‍⚕👨🏻‍⚕🏳️‍🌇🏳️‍🌆🏳️‍🌅🏴󠁧󠁢󠁥󠁮󠁧󠁿🏴󠁧<#selection-caret>"

Dump of markup 14:
| "
🤙🏻👨‍⚕️👨🏻‍⚕️👨‍⚕👨🏻‍⚕🏳️‍🌇🏳️‍🌆🏳️‍🌅🏴󠁧󠁢󠁥󠁮󠁧󠁿🏴<#selection-caret>"

Dump of markup 15:
| "
🤙🏻👨‍⚕️👨🏻‍⚕️👨‍⚕👨🏻‍⚕🏳️‍🌇🏳️‍🌆🏳️‍🌅🏴󠁧󠁢󠁥󠁮󠁧󠁿<#selection-caret>"

Dump of markup 16:
| "
🤙🏻👨‍⚕️👨🏻‍⚕️👨‍⚕👨🏻‍⚕🏳️‍🌇🏳️‍🌆🏳️‍🌅🏴󠁧󠁢󠁥󠁮󠁧<#selection-caret>"

Dump of markup 17:
| "
🤙🏻👨‍⚕️👨🏻‍⚕️👨‍⚕👨🏻‍⚕🏳️‍🌇🏳️‍🌆🏳️‍🌅🏴󠁧󠁢󠁥󠁮<#selection-caret>"

Dump of markup 18:
| "
🤙🏻👨‍⚕️👨🏻‍⚕️👨‍⚕👨🏻‍⚕🏳️‍🌇🏳️‍🌆🏳️‍🌅🏴󠁧󠁢󠁥<#selection-caret>"

Dump of markup 19:
| "
🤙🏻👨‍⚕️👨🏻‍⚕️👨‍⚕👨🏻‍⚕🏳️‍🌇🏳️‍🌆🏳️‍🌅🏴󠁧󠁢<#selection-caret>"

Dump of markup 20:
| "
🤙🏻👨‍⚕️👨🏻‍⚕️👨‍⚕👨🏻‍⚕🏳️‍🌇🏳️‍🌆🏳️‍🌅🏴󠁧<#selection-caret>"

Dump of markup 21:
| "
🤙🏻👨‍⚕️👨🏻‍⚕️👨‍⚕👨🏻‍⚕🏳️‍🌇🏳️‍🌆🏳️‍🌅🏴<#selection-caret>"

Dump of markup 22:
| "
🤙🏻👨‍⚕️👨🏻‍⚕️👨‍⚕👨🏻‍⚕🏳️‍🌇🏳️‍🌆🏳️‍🌅<#selection-caret>"

Dump of markup 23:
| "
🤙🏻👨‍⚕️👨🏻‍⚕️👨‍⚕👨🏻‍⚕🏳️‍🌇🏳️‍🌆🏳️‍<#selection-caret>"

Dump of markup 24:
| "
🤙🏻👨‍⚕️👨🏻‍⚕️👨‍⚕👨🏻‍⚕🏳️‍🌇🏳️‍🌆🏳️<#selection-caret>"

Dump of markup 25:
| "
🤙🏻👨‍⚕️👨🏻‍⚕️👨‍⚕👨🏻‍⚕🏳️‍🌇🏳️‍🌆<#selection-caret>"

Dump of markup 26:
| "
🤙🏻👨‍⚕️👨🏻‍⚕️👨‍⚕👨🏻‍⚕🏳️‍🌇🏳️‍<#selection-caret>"

Dump of markup 27:
| "
🤙🏻👨‍⚕️👨🏻‍⚕️👨‍⚕👨🏻‍⚕🏳️‍🌇🏳️<#selection-caret>"

Dump of markup 28:
| "
🤙🏻👨‍⚕️👨🏻‍⚕️👨‍⚕👨🏻‍⚕🏳️‍🌇<#selection-caret>"

Dump of markup 29:
| "
🤙🏻👨‍⚕️👨🏻‍⚕️👨‍⚕👨🏻‍⚕🏳️‍<#selection-caret>"

Dump of markup 30:
| "
🤙🏻👨‍⚕️👨🏻‍⚕️👨‍⚕👨🏻‍⚕🏳️<#selection-caret>"

Dump of markup 31:
| "
🤙🏻👨‍⚕️👨🏻‍⚕️👨‍⚕👨🏻‍⚕<#selection-caret>"

Dump of markup 32:
| "
🤙🏻👨‍⚕️👨🏻‍⚕️👨‍⚕<#selection-caret>"

Dump of markup 33:
| "
🤙🏻👨‍⚕️👨🏻‍⚕️<#selection-caret>"

Dump of markup 34:
| "
🤙🏻👨‍⚕️<#selection-caret>"

Dump of markup 35:
| "
🤙🏻<#selection-caret>"

Dump of markup 36:
| <#selection-caret>
| <br>
